Trait asset_store::AssetStore

pub trait AssetStore<E> {
    fn load(&mut self, path: &str);
    fn is_loaded(&mut self, path: &str) -> Result<bool, E>;
    fn unload(&mut self, path: &str);
    fn unload_everything(&mut self);
    fn fetch<'a>(&'a mut self, path: &str) -> Result<Option<&'a [u8]>, E>;
    fn fetch_block<'a>(&'a mut self, path: &str) -> Result<&'a [u8], E>;

    fn load_all<'a, I: Iterator<&'a str>>(&mut self, paths: I) { ... }
    fn all_loaded<'a, I: Iterator<&'a str>>(&mut self, paths: I) -> Result<bool, Vec<(&'a str, E)>> { ... }
    fn unload_all<'a, I: Iterator<&'a str>>(&mut self, paths: I) { ... }
}

Required Methods

fn load(&mut self, path: &str)

Tell the asset store to begin loading a resource.

fn is_loaded(&mut self, path: &str) -> Result<bool, E>

Check to see if a resource has been loaded or not.

fn unload(&mut self, path: &str)

Remove this resouce from this asset store if it is loaded.

fn unload_everything(&mut self)

Remove every resouce from this asset store

fn fetch<'a>(&'a mut self, path: &str) -> Result<Option<&'a [u8]>, E>

Try to fetch a resource. If the resource is fully loaded, returns Ok(Some(resource)) If the resource has not been loaded, returns Ok(None) If the resource failed to load, returns Err(e)

fn fetch_block<'a>(&'a mut self, path: &str) -> Result<&'a [u8], E>

Try to fetch a resource. If the resource has not been loaded yet, block until it is loaded.

Provided Methods

fn load_all<'a, I: Iterator<&'a str>>(&mut self, paths: I)

Tell the asset store to begin loading all resources.

fn all_loaded<'a, I: Iterator<&'a str>>(&mut self, paths: I) -> Result<bool, Vec<(&'a str, E)>>

Check to see if everything has been loaded.

fn unload_all<'a, I: Iterator<&'a str>>(&mut self, paths: I)

Remove all these resouces from this asset store if they are loaded.